Moved stm sensor to libs
