+++ MXChip/AZ3166/app/CMakeLists.txt
@@ -7,16 +7,7 @@ set(SOURCES
     startup/system_stm32f4xx.c
     startup/startup_stm32f412rx.s
     startup/tx_initialize_low_level.S
-    
-    sensor/lps22hb_read_data_polling.c
-    sensor/lps22hb_reg.c    
-    sensor/hts221_read_data_polling.c
-    sensor/hts221_reg.c  
-    sensor/lsm6dsl_read_data_polling.c
-    sensor/lsm6dsl_reg.c       
-    sensor/lis2mdl_read_data_polling.c
-    sensor/lis2mdl_reg.c 
-    
+   
     azure_iothub.c
     board_init.c
     console.c
@@ -37,8 +28,7 @@ target_link_options(${PROJECT_NAME}
     PUBLIC 
         -T${LINKER_SCRIPT} -Wl,-Map=${PROJECT_NAME}.map)
 
-target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}
-                           ${PROJECT_NAME} PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/sensor/")
+target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
 target_link_libraries(${PROJECT_NAME}
     azrtos::threadx
