+++ MXChip/AZ3166/lib/stm32cubef4/Drivers/stm_sensor/Inc/hts221_reg.h
