+++ MXChip/AZ3166/lib/stm32cubef4/Drivers/stm_sensor/Src/hts221_read_data_polling.c
