+++ MXChip/AZ3166/lib/stm32cubef4/Drivers/stm_sensor/Src/lis2mdl_reg.c
